+++ src/components/TodoEditComponents/Head/index.tsx
@@ -1,22 +1,25 @@
 import React, { FC, Component } from 'react';
 import styled from 'styled-components/native';
-import CustomButton from './CustomButton';
+import {BackButton, DeleteButton, ColorPicker} from './HeadComponents';
 import { connect } from 'react-redux';
 
 import Colors from '../../../constans/Colors';
+import { Picker } from 'react-native';
+import NoteColors from '../../../constans/NoteColors';
 
 const HeadContainer = styled.View`
     width: 100%;
-    backgroundColor: ${Colors.mediumGray};
+    backgroundColor: ${Colors.darkGray};
     border: solid 1px ${Colors.black};
+    flexDirection: row;
 `;
 
 const Head: FC = () => {
     return (
         <HeadContainer>
-            <CustomButton text={'←'} onPress={() => undefined}/>
-
-            <CustomButton text={'☓'} onPress={() => undefined}/>
+            <BackButton onPress={() => undefined}/>
+            <ColorPicker selectedColor={NoteColors.red} />
+            <DeleteButton onPress={() => undefined}/>
         </HeadContainer>
     );
 };